refactor: migrate workflow pages from packages/features to apps/web/modules
- Move packages/features/ee/workflows/pages/index.tsx to apps/web/modules/workflows/views/workflows-listing-view.tsx
- Move packages/features/ee/workflows/pages/workflow.tsx to apps/web/modules/workflows/views/workflow-view.tsx
- Update imports in apps/web/app/(use-page-wrapper)/workflows/ to use new paths
- Remove unused Filter component and clean up imports
- This continues the work from PR #25859 to eradicate trpc imports in packages/features
