+++ apps/web/app/(use-page-wrapper)/workflows/page.tsx
@@ -3,7 +3,7 @@
 // import { buildLegacyRequest } from "@lib/buildLegacyCtx";
 // import { getTeamsFiltersFromQuery } from "@calcom/features/filters/lib/getTeamsFiltersFromQuery";
 // import { WorkflowRepository } from "@calcom/features/ee/workflows/repositories/WorkflowRepository";
-import LegacyPage from "@calcom/features/ee/workflows/pages/index";
+import LegacyPage from "~/workflows/views/workflows-listing-view";
 
 const Page = async () => {
   // const session = await getServerSession({ req: buildLegacyRequest(await headers(), await cookies()) });
